refactor: reorganize components and services structure, update imports to match new paths
