+++ src/components/layout/Header/Header.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { useSession } from 'next-auth/react';
-import Header from './Header';
+import Header from '@components/layout/Header/Header';
 import { vi, describe, beforeAll, test, expect } from 'vitest';
 
 vi.mock('next-auth/react', () => {
