+++ src/components/ui/Form.tsx
@@ -11,7 +11,7 @@ import {
 } from 'react-hook-form';
 
 import { cn } from '@/lib/utils';
-import { Label } from '@/components/ui/Label';
+import { Label } from '@/components/ui/label';
 
 const Form = FormProvider;
