+++ src/features/generate/components/generate-playlist-form/playlist-search-artists-form/artist-search-result.tsx
